requirements.txt
Flask==3.0.2
gunicorn==21.2.0
requests==2.32.3
python-binance==1.0.19

